Change to STRICT_RC calculation of active best Q.
This on average improves rate accuracy and metrics when this flag
is set.

Low Res: -6.683	-9.350	-4.811	-7.751	-9.960
UGC 360: -4.748	-8.286	-3.154	-5.952	-7.703
Mid Res: -3.122	-6.052	-0.598	-4.436	-5.823
Hd Res:  -5.381	-6.944	-2.542	-5.649	-7.736

Note however that there are a few clips where metrics are much worse
all be it with big improvements to rate control accuracy across the
clip.

This is exemplified by the screen res set where undershoot is
generally reduced but metrics are down.  This can be followed up on
a separate patch and if need be screen content can be handled
differently.

Change-Id: I704f4712fdacc09289a3f5fec00da48edb4906b9