rtc: Add dynamic resize mode, to 1 pass CBR
Make it work for 1 pass CBR mode.
Add a DYNAMIC_RESIZE mode and allow for 3/4 and
1/2 resize factors, where the resize logic is based
on the rate control.

Also force cyclic_refresh off on the resize trigger
frame. And disable golden and altref when the
resolution of the reference frame is different from
the source (to avoid slowness in scale_references).

Re-enabled/update two unittests for dynamic resize.

Change-Id: I10edf3ca9c2e90ba8ed8bbfac598c18aa442a5bf