Validate the has_val field of arg_def in arg_match
Assert that the has_val field of arg_def is one of the three valid
values (0, 1, -1) in arg_match(). Since arg_match() is invoked on every
arg_def, this allows us to detect incorrect values of has_val in debug
builds or in builds with assertions enabled.

Include "common/tools_common.h" for die() and remove the manual
declaration of die().

Change-Id: I9ed87f61075c9f86790c9f5d2e872343fd46fa45