Modify the assertion in temporal filter intrinsics
This CL is a manual backport of commit
1fe2def01070c5583d44023770c950060e05f3b8 to the applejack branch (used
for libaom 2.0.x releases). I inspected the AVX2 and SSE2 versions of
the apply_temporal_filter_planewise() function on the applejack branch.
Althugh they do require block_width to be 32 or 16, they only require
block_height to be greater than or equal to 3 or 4. So the assertion can
be relaxed.

History of this bug:
- First fix: https://aomedia-review.googlesource.com/c/aom/+/109322
- Second fix: https://aomedia-review.googlesource.com/c/aom/+/117843

BUG=aomedia:2643

Change-Id: I28093c54a2606e0118bdc3693a3bc6b5aee95689