Initialize coeff, mode and mv costs
When cost update frequency was set to tile level,
coeff_costs, mode_costs and mv costs were not initialized.
This resulted in run to run mismatch in the multi-thread
scenario with row-mt=1. Also, when cost update frequency
was set to off, coeff_costs and mode_costs were initialized
to zero. This patch correctly initializes these costs.

Change-Id: I4193185058be045b9711d2bb2df77dbf5ba8634b