Merge the InitializeConfig() and SetMode() methods
Merge the InitializeConfig() and SetMode() methods of the EncoderTest
class. Subclasses of EncoderTest always call InitializeConfig() and
SetMode() (in that order) in their SetUp() methods. So these two methods
can be combined. This allows us to pass a 'usage' argument corresponding
to the TestMode enum to codec_->DefaultEncoderConfig().

Remove the assignments that set cfg_.g_lag_in_frames and
cfg_.rc_end_usage to their current (default) values.

BUG=aomedia:2966

Change-Id: I2335f91fbe49c7ab46e4a0c8a73d4fdbfeec3eb2