Add an emms instruction to aom_subtract_block
This commit adds an emms instruction at the end of MMX assembly for
aom_subtract_block to properly clear the register state.

Note that there is a slight stats difference, indicating that there is a
missing aom_clear_system_state somewhere in the codebase.

This also resolves a mismatch between x86 build and x86-64 build. See
the bug tracker for more details.

BUG=aomedia:2968

STATS_CHANGED

Change-Id: I22b0269017007a5d35ad19cd7e15a838d5ebd326