rtc: Redefine the speed feature hybrid_intra_pickmode
Define it to be 0/off: all blocks use nonrd pick intra,
1/on: hybrid is used where small blocks go through the
rd pick intra.

Current setting is 1/on for all speeds, so no change in behavior.

If set to 0, then speedup is ~4x/8x relative to speed 8/7,
bdrate loss for all key frames: ~15%/18% relative to speed 8/7.

Change threshold in RTEndToEndTest, this was needed from:f38266

Change-Id: Idf48fd5fe650c48e92435699c0b5df2599ce8587