CMakeLists.txt: explicitly set CMP0118 to NEW
this quiets a developer warning in 3.20.0:

Policy CMP0118 is not set: The GENERATED source file property is now
visible in all directories

and as part of the policy explicitly sets the value to TRUE to quiet:

CMake Warning (dev) at build/cmake/aom_optimization.cmake:241 (set_property):
  Unsetting the 'GENERATED' property is not allowed under CMP0118!

Call Stack (most recent call first):
  CMakeLists.txt:237 (add_rtcd_build_step)
This warning is for project developers.  Use -Wno-dev to suppress it.

Bug: aomedia:3015
Change-Id: Id713c39e5fd16908805b95b4963316a6d0a12154