External partition: Separate multi decisions in one stage
This cl has a few purposes:
(1). Separate the partition features and call functions when
there are more than one partition decision is made in a particular
partition stage, for example, after partition none and split.
(2). Add unique ids for these separated stages accordingly.
(3). Add a return status for external partition, so that we could
use it for tests.
(4). Rework functions where we collect partition features and call
external partition functions. Make it align with the original
ml model.

Change-Id: Ib542ab4d4245cfb25c375c9496049ad2bff0c043