Fix DecodeScalabilityTest failure in realtime only
Fix DecodeScalabilityTest failure in the CONFIG_REALTIME_ONLY builds.
The three test vectors used by DecodeScalabilityTest all use loop
restoration, which is not supported in the CONFIG_REALTIME_ONLY builds.

Add a new test vector that does not use scalability so that we can still
run DecodeScalabilityTest on that test vector in the
CONFIG_REALTIME_ONLY builds. Ideally we should create new test vectors
that use scalability but not loop restoration.

BUG=aomedia:2993,aomedia:3077

Change-Id: Idf85622bae14513ff8f78e954b6ad72bd7c1c669