Fix tx_type setting in a corner case
In tx type search, for blocks with very low residual variance, the
search was bypassed. In this situation, tx_type_map needed to be
set accordingly, otherwise, the encoder failed as reported in the
issue. This CL fixed the bug. This seemed to only happen at high
speed presets since the related speed feature was on for speed >= 5.

Borg test with --enable-tx64=1 at speed 5 and 6 showed negligible
impact (< 0.01%) on coding efficiency.

BUG=aomedia:3123

STATS_CHANGED

Change-Id: I21110c21ccf973368258f92464c813915a958902