ExtPart: encode with ml decision recursively
This approach is similar to the baseline DFS partition search.
Each time the ML model makes one decision for the current block,
then the encoder encodes and send information back to the external
model.
Such process starts from the super block and proceeds recursively
to smaller blocks when split happens.

Change-Id: Id81a3b747938735567d833625efb986821a72596