ExtPart: unit tests for whole partition tree
Add a unit test that compares psnr values to two encoding runs.

The first run is the baseline encoder, where we restrict the max
and min coding blocks to 4x4.

In the second run, we read partition decisions from the a toy model,
which provides the whole partition decision tree for the superblock.
And the decision tree is exactly all splits down to 4x4 block.

We expect these two runs to return the same psnr value.

Change-Id: Iec81f83164ad6327aa41278dc9822b91f11610a9