Allintra: Reduce border_in_pixels for frame buffers
For allintra encoding mode, inter-frame motion search is not
applicable and the intraBC motion vectors are restricted within the
tile boundaries. Hence, the border value used in the allocation of
frame buffer is reduced from 160 pixels to 64 pixels in case of
allintra encoding mode.

For AVIF image encode with speed = 9,

             HEAP Memory reduction(%)
Resolution   threads=1    threads=4
640x360        9.00         5.64
768x512        6.92         4.80
832x480        7.20         5.16
1280x720       5.03         4.10

For threads=4, an average encode time reduction of ~3.47% is
observed for 360p-720p resolutions.

HEAP memory reduction was measured using the following command.
$valgrind --tool=massif ./avifenc ...

Change-Id: Ibf690032b4c86a07f7751889de36b21e94544f1c