Add variance match weighting to all intra rd.
Add a weighting to the rd choice based on how well the reconstruction
variance matches the source variance. This is designed to bias against
mode choices that cause a big change (mainly drop) in spatial complexity
and is intended to improve perceived sharpness. It does have a negative
impact on metrics (especially psnr) which tend to prefer a smoothed
outcome. In most cases tested so far this drop is modest.

Change-Id: Ia3d69ae34d68ea0c20bf7fa763a6d9ab104aa169