ExtPart: Unit test for recursive ml partition
We compare the psnr values to two encoding runs.

The first run is encoding with the baseline, by restricting block sizes
to the super block size.

The second run is partition decisions from a toy model that provides
matching partition decisions as the first run.

The psnr values differ by a tiny amount (less than 0.001 db).

Change-Id: I677e616f284c744a9d7735f5c4777a57de0ac816