Cap no. of txfm searches for a prediction mode
The number of inter mode candidates for transform search is capped
beyond a threshold as per the prediction mode, up to 2 for NEARESTMV
modes and up to 1 for other modes. This is applicable for speed >= 3.

          Instruction Count       BD-Rate Loss(%)
cpu-used    Reduction(%)     avg.psnr  ovr.psnr   ssim
   3          1.149           0.0219    0.0116   -0.0102

For speed >= 4, no significant impact on speed or quality.

STATS_CHANGED

Change-Id: I686c8b862f47761897e4fd9bd7ea7e21a9d20e1f