Add the aom_tools_ prefix to our warn() function
Add the aom_tools_ prefix to the warn() function in
common/tools_common.h.

It turns out that glibc also has a warn() function with the same
prototype. So on Linux systems that use glibc, it is impossible to
detect unresolved references to our warn() function in the libaom shared
library. Renaming our warn() function allows us to detect this error on
glibc systems.

BUG=aomedia:3139

Change-Id: I4a0da9b2ba211c2b0a8fdb5a57cd5aa13009d830