rtc: Enable the use of golden ref for speed < 7
The golden ref was disabled to avoid encoder speed decrease in
https://aomedia-review.googlesource.com/c/aom/+/142041
This was re-evaluated for rt speed 5 and 6. The coding gain was
significant and outweighed the speed drop. Thus it was enabled.
Here were borg test results.
          avg_psnr:  ovr_psnr:  ssim:  encoder speed:
speed 5:
rtc_derf:  -3.041    -3.112    -3.377    -11.518
rtc:       -2.539    -2.295    -2.519    -9.124
speed 6:
rtc_derf:  -2.993    -3.097    -3.363    -9.622
rtc:       -2.459    -2.251    -2.471    -8.235

STATS_CHANGED for rt speed < 7

Change-Id: Icaa3135ab3922e32f0b4d09caf0c5571d4c8e22b