Only compute tpl stats when gf_frame_index = 0
Previously in the first GOP, we will compute tpl stats
at key frame with gf_frame_index = 1 and ARF frame with
gf_frame_index = 0. In this CL, we merged the to runs
of TPL stats computation.

We see mostly neutral results except for midres there is 0.1% drop for
ssim and for ugc360p, hdres and lowres there is 0.1% gain for ssim

q mode --lag-in-frames=48 --limit=150

(negative is better)
       avg_psnr ovr_psnr ssim
ugc360p 0.018    0.001   -0.104
lowres  0.020    0.041   -0.094
midres -0.039    0.047    0.138
hdres  -0.078   -0.084   -0.063
hdres2 -0.062   -0.070   -0.131

BUG=aomedia:3144

STATS_CHANGED

Change-Id: I3216d86c54f721019110b58f9c7a1befbdeea9b5