Auto select noise synthesis level for all intra
Add code to estimate the noise level in a still image and choose
a noise synthesis level.

The auto selection of the noise synthesis level will only occur where
--allintra is set on the command line and --denoise-noise-level > 0.

Also on the command line:

if --enable-dnl-denoising=1 then the source image will also be filtered.
If --enable-dnl-denoising=0  then the source image will not be filtered.

Filtering of the source image is much more likely to cause some loss
fine detail if a strong filter level is chosen and is not recommended.

Adjustment from previous patch increases filtering on some images
where subjective evaluation said it was too low and reduced it on
some high complexity images where it was too high (by lowering the
edge threshold).

Change-Id: I90c99c3c12ddb8f5155f0ad7d87c35767ffccb38