rtc: Reduce memory allocation for mode info
In real time mode, min_partition_size used is BLOCK_8X8. So, we
don't need to have mode info (mi) in 4x4 precision. In this CL,
we allocate mi in 8x8 precision for rtc encoder, which largely
reduces encoder memory usage for all rt speeds (speed 5 - 10).

In average, the encoder memory reduction is 15.3% for 720p, 9%
for 360p, and 4.6% for 180p. No stats change is introduced.

Change-Id: I66c27b6713a501c39799ad1d10dba366e3d91548