rtc: Speed feature to switch to half_pel at block level
Add speed feature, set at speed >= 10, to allow for
switching to half_pel for newmv at coding block level.
Switch is based on the size of the integer motion vector,
along with bsize and avg_frame_low_motion. Lower threshold
is used for lower resolutions.

avg. bdrate loss ~0.6/0.7 on rtc/rtc_derf_set, max loss is 1.7
speedup ~2-3% on some vga and qvga clips.
This is for speed 10, with cdf-update-mode=2.

Change-Id: I38201e2a9d5ac8336045b2b58168686b541d2933