Only use av1_temporal_filter when filter is needed
Move the logics of checking second arf out of av1_temporal_filter.

Move the showable_frame logics out of av1_temporal_filter.

Before, we assumed av1_temporal_filter will be called when
the filtered frame is about to be coded, but it's not the
case anymore. The temporal filter can be called at any time.
Therefore, we need to move those logics out.

BUG=aomedia:3144

Change-Id: I4ed5ded65a1d559abc66638b0a4ee0c7948cac6b