rtc: Fix assert in nonrd pickmode
The assert checks that one of the referenes
(last, golden, altref) is used for inter mode selection.
Condition the assert on svc mode, as for svc the user
is responsible for setting the references, and there
could cases where no reference is intentionally set.

Also temporarily remove the assert(target_bandwidth > 0)
in the update_layer_context_change(), as this is causing
some test failures that need to fixed first.

Change-Id: Ic8504a2429c1eb66a5e19a951b778a85f38a0350